pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-trader"
version = "0.1.0"
description = "Automated cryptocurrency trading bot for Binance"
requires-python = ">=3.11"
dependencies = [
    "python-binance>=1.0.19",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    "scikit-learn>=1.4.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.10.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
]
ml = [
    "torch>=2.2.0",
    "xgboost>=2.0.0",
    "optuna>=3.5.0",
    "joblib>=1.3.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
auto-trader = "src.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
disallow_untyped_defs = true
